refactor(cli): merge project-context helpers into project-root workflow module
Agent-Logs-Url: https://github.com/SimulatorLife/GMLoop/sessions/1b26cf70-e0f4-4463-87d7-c12feb776b1e
