+++ src/cli/src/workflow/index.ts
@@ -1,4 +1,3 @@
 export * from "./display-path.js";
 export * from "./path-filter.js";
-export * from "./project-context.js";
 export * from "./project-root.js";
